Add file size checking script
- Introduce a new script `check_file_sizes.sh` to evaluate the size of files in a directory based on the number of lines.
- Implement functionality to categorize files as "OK", "Acceptable", or "BAD" based on their line count, with corresponding color-coded outputs.
- Include options to specify file extensions to check, exclude test files, and process all text files.
- Add command-line argument parsing and help documentation for user guidance.

This commit enhances the development workflow by providing a tool for maintaining file size standards in the codebase.
