+++ internal/diagfmt/ast_expr.go
@@ -1,0 +1,311 @@
+package diagfmt
+
+import (
+	"fmt"
+	"strings"
+	"surge/internal/ast"
+	"surge/internal/source"
+)
+
+const exprInlineMaxDepth = 32
+
+func formatExprSummary(builder *ast.Builder, exprID ast.ExprID) string {
+	if !exprID.IsValid() {
+		return "<none>"
+	}
+	inline := formatExprInlineDepth(builder, exprID, 0)
+	if inline == "" {
+		inline = "<invalid>"
+	}
+	return fmt.Sprintf("expr#%d: %s", uint32(exprID), inline)
+}
+
+func formatExprInline(builder *ast.Builder, exprID ast.ExprID) string {
+	return formatExprInlineDepth(builder, exprID, 0)
+}
+
+func formatExprInlineDepth(builder *ast.Builder, exprID ast.ExprID, depth int) string {
+	if !exprID.IsValid() {
+		return "<none>"
+	}
+	if builder == nil || builder.Exprs == nil {
+		return "<invalid>"
+	}
+	if depth >= exprInlineMaxDepth {
+		return "..."
+	}
+
+	expr := builder.Exprs.Get(exprID)
+	if expr == nil {
+		return "<invalid>"
+	}
+
+	switch expr.Kind {
+	case ast.ExprIdent:
+		data, ok := builder.Exprs.Ident(exprID)
+		if !ok {
+			return "<invalid-ident>"
+		}
+		if builder.StringsInterner == nil || data.Name == source.NoStringID {
+			return "<ident>"
+		}
+		return builder.StringsInterner.MustLookup(data.Name)
+	case ast.ExprLit:
+		data, ok := builder.Exprs.Literal(exprID)
+		if !ok {
+			return "<invalid-literal>"
+		}
+		switch data.Kind {
+		case ast.ExprLitTrue:
+			return "true"
+		case ast.ExprLitFalse:
+			return "false"
+		case ast.ExprLitNothing:
+			return "nothing"
+		default:
+			if builder.StringsInterner != nil && data.Value != source.NoStringID {
+				return builder.StringsInterner.MustLookup(data.Value)
+			}
+			return "<literal>"
+		}
+	case ast.ExprUnary:
+		data, ok := builder.Exprs.Unary(exprID)
+		if !ok {
+			return "<invalid-unary>"
+		}
+		operand := formatExprInlineDepth(builder, data.Operand, depth+1)
+		operand = wrapExprIfNeeded(builder, data.Operand, operand)
+		return formatUnaryOpString(data.Op, operand)
+	case ast.ExprBinary:
+		data, ok := builder.Exprs.Binary(exprID)
+		if !ok {
+			return "<invalid-binary>"
+		}
+		left := formatExprInlineDepth(builder, data.Left, depth+1)
+		right := formatExprInlineDepth(builder, data.Right, depth+1)
+		left = wrapExprIfNeeded(builder, data.Left, left)
+		right = wrapExprIfNeeded(builder, data.Right, right)
+		op := formatBinaryOpString(data.Op)
+		return fmt.Sprintf("(%s %s %s)", left, op, right)
+	case ast.ExprCall:
+		data, ok := builder.Exprs.Call(exprID)
+		if !ok {
+			return "<invalid-call>"
+		}
+		target := formatExprInlineDepth(builder, data.Target, depth+1)
+		target = wrapExprIfNeeded(builder, data.Target, target)
+		args := make([]string, 0, len(data.Args))
+		for _, arg := range data.Args {
+			args = append(args, formatExprInlineDepth(builder, arg, depth+1))
+		}
+		return fmt.Sprintf("%s(%s)", target, strings.Join(args, ", "))
+	case ast.ExprIndex:
+		data, ok := builder.Exprs.Index(exprID)
+		if !ok {
+			return "<invalid-index>"
+		}
+		target := formatExprInlineDepth(builder, data.Target, depth+1)
+		target = wrapExprIfNeeded(builder, data.Target, target)
+		index := formatExprInlineDepth(builder, data.Index, depth+1)
+		return fmt.Sprintf("%s[%s]", target, index)
+	case ast.ExprMember:
+		data, ok := builder.Exprs.Member(exprID)
+		if !ok {
+			return "<invalid-member>"
+		}
+		target := formatExprInlineDepth(builder, data.Target, depth+1)
+		target = wrapExprIfNeeded(builder, data.Target, target)
+		field := "<field>"
+		if builder.StringsInterner != nil && data.Field != source.NoStringID {
+			field = builder.StringsInterner.MustLookup(data.Field)
+		}
+		return fmt.Sprintf("%s.%s", target, field)
+	case ast.ExprGroup:
+		data, ok := builder.Exprs.Group(exprID)
+		if !ok {
+			return "<invalid-group>"
+		}
+		inner := formatExprInlineDepth(builder, data.Inner, depth+1)
+		return fmt.Sprintf("(%s)", inner)
+	case ast.ExprTuple:
+		data, ok := builder.Exprs.Tuple(exprID)
+		if !ok {
+			return "<invalid-tuple>"
+		}
+		if len(data.Elements) == 0 {
+			return "()"
+		}
+		elems := make([]string, 0, len(data.Elements))
+		for _, elem := range data.Elements {
+			elems = append(elems, formatExprInlineDepth(builder, elem, depth+1))
+		}
+		if len(data.Elements) == 1 {
+			return fmt.Sprintf("(%s,)", elems[0])
+		}
+		return fmt.Sprintf("(%s)", strings.Join(elems, ", "))
+	case ast.ExprCast:
+		data, ok := builder.Exprs.Cast(exprID)
+		if !ok {
+			return "<invalid-cast>"
+		}
+		value := formatExprInlineDepth(builder, data.Value, depth+1)
+		value = wrapExprIfNeeded(builder, data.Value, value)
+		typ := formatTypeExprInline(builder, data.Type)
+		return fmt.Sprintf("%s to %s", value, typ)
+	default:
+		return fmt.Sprintf("<%s>", formatExprKind(expr.Kind))
+	}
+}
+
+func wrapExprIfNeeded(builder *ast.Builder, exprID ast.ExprID, rendered string) string {
+	if !exprID.IsValid() {
+		return rendered
+	}
+	if builder == nil || builder.Exprs == nil {
+		return rendered
+	}
+	expr := builder.Exprs.Get(exprID)
+	if expr == nil {
+		return rendered
+	}
+
+	switch expr.Kind {
+	case ast.ExprBinary, ast.ExprTernary, ast.ExprCompare, ast.ExprCast:
+		return "(" + rendered + ")"
+	default:
+		return rendered
+	}
+}
+
+func formatUnaryOpString(op ast.ExprUnaryOp, operand string) string {
+	switch op {
+	case ast.ExprUnaryPlus:
+		return "+" + operand
+	case ast.ExprUnaryMinus:
+		return "-" + operand
+	case ast.ExprUnaryNot:
+		return "!" + operand
+	case ast.ExprUnaryDeref:
+		return "*" + operand
+	case ast.ExprUnaryRef:
+		return "&" + operand
+	case ast.ExprUnaryRefMut:
+		return "&mut " + operand
+	case ast.ExprUnaryOwn:
+		return "own " + operand
+	case ast.ExprUnaryAwait:
+		return "await " + operand
+	default:
+		return fmt.Sprintf("<unary %d> %s", op, operand)
+	}
+}
+
+func formatBinaryOpString(op ast.ExprBinaryOp) string {
+	switch op {
+	case ast.ExprBinaryAdd:
+		return "+"
+	case ast.ExprBinarySub:
+		return "-"
+	case ast.ExprBinaryMul:
+		return "*"
+	case ast.ExprBinaryDiv:
+		return "/"
+	case ast.ExprBinaryMod:
+		return "%"
+	case ast.ExprBinaryBitAnd:
+		return "&"
+	case ast.ExprBinaryBitOr:
+		return "|"
+	case ast.ExprBinaryBitXor:
+		return "^"
+	case ast.ExprBinaryShiftLeft:
+		return "<<"
+	case ast.ExprBinaryShiftRight:
+		return ">>"
+	case ast.ExprBinaryLogicalAnd:
+		return "&&"
+	case ast.ExprBinaryLogicalOr:
+		return "||"
+	case ast.ExprBinaryEq:
+		return "=="
+	case ast.ExprBinaryNotEq:
+		return "!="
+	case ast.ExprBinaryLess:
+		return "<"
+	case ast.ExprBinaryLessEq:
+		return "<="
+	case ast.ExprBinaryGreater:
+		return ">"
+	case ast.ExprBinaryGreaterEq:
+		return ">="
+	case ast.ExprBinaryAssign:
+		return "="
+	case ast.ExprBinaryAddAssign:
+		return "+="
+	case ast.ExprBinarySubAssign:
+		return "-="
+	case ast.ExprBinaryMulAssign:
+		return "*="
+	case ast.ExprBinaryDivAssign:
+		return "/="
+	case ast.ExprBinaryModAssign:
+		return "%="
+	case ast.ExprBinaryBitAndAssign:
+		return "&="
+	case ast.ExprBinaryBitOrAssign:
+		return "|="
+	case ast.ExprBinaryBitXorAssign:
+		return "^="
+	case ast.ExprBinaryShlAssign:
+		return "<<="
+	case ast.ExprBinaryShrAssign:
+		return ">>="
+	case ast.ExprBinaryNullCoalescing:
+		return "??"
+	case ast.ExprBinaryRange:
+		return ".."
+	case ast.ExprBinaryRangeInclusive:
+		return "..="
+	case ast.ExprBinaryIs:
+		return "is"
+	default:
+		return fmt.Sprintf("op%d", op)
+	}
+}
+
+func formatExprKind(kind ast.ExprKind) string {
+	switch kind {
+	case ast.ExprIdent:
+		return "Ident"
+	case ast.ExprLit:
+		return "Literal"
+	case ast.ExprCall:
+		return "Call"
+	case ast.ExprBinary:
+		return "Binary"
+	case ast.ExprUnary:
+		return "Unary"
+	case ast.ExprCast:
+		return "Cast"
+	case ast.ExprGroup:
+		return "Group"
+	case ast.ExprTuple:
+		return "Tuple"
+	case ast.ExprIndex:
+		return "Index"
+	case ast.ExprMember:
+		return "Member"
+	case ast.ExprTernary:
+		return "Ternary"
+	case ast.ExprAwait:
+		return "Await"
+	case ast.ExprSignal:
+		return "Signal"
+	case ast.ExprParallel:
+		return "Parallel"
+	case ast.ExprCompare:
+		return "Compare"
+	default:
+		return fmt.Sprintf("ExprKind(%d)", kind)
+	}
+}
