+++ internal/diagfmt/ast_item.go
@@ -1,0 +1,241 @@
+package diagfmt
+
+import (
+	"fmt"
+	"io"
+	"surge/internal/ast"
+	"surge/internal/source"
+)
+
+func formatItemPretty(w io.Writer, builder *ast.Builder, itemID ast.ItemID, fs *source.FileSet, prefix string) error {
+	item := builder.Items.Get(itemID)
+	if item == nil {
+		fmt.Fprintf(w, "nil item\n")
+		return nil
+	}
+
+	kindStr := formatItemKind(item.Kind)
+	fmt.Fprintf(w, "%s (span: %s)\n", kindStr, formatSpan(item.Span, fs))
+
+	// Handle special items with payload
+	switch item.Kind {
+	case ast.ItemImport:
+		if importItem, ok := builder.Items.Import(itemID); ok {
+			// Count how many fields we have to determine which one is last
+			hasAlias := importItem.ModuleAlias != 0
+			hasOne := importItem.HasOne
+			hasGroup := len(importItem.Group) > 0
+
+			fieldsCount := 1 // always have Module
+			if hasAlias {
+				fieldsCount++
+			}
+			if hasOne {
+				fieldsCount++
+			}
+			if hasGroup {
+				fieldsCount++
+			}
+
+			currentField := 0
+
+			// Module is always first
+			currentField++
+			modulePrefix := "├─"
+			if currentField == fieldsCount {
+				modulePrefix = "└─"
+			}
+			fmt.Fprintf(w, "%s%s Module: ", prefix, modulePrefix)
+			for i, stringID := range importItem.Module {
+				if i > 0 {
+					fmt.Fprintf(w, "::")
+				}
+				fmt.Fprintf(w, "%s", builder.StringsInterner.MustLookup(stringID))
+			}
+			fmt.Fprintf(w, "\n")
+
+			if hasAlias {
+				currentField++
+				aliasPrefix := "├─"
+				if currentField == fieldsCount {
+					aliasPrefix = "└─"
+				}
+				fmt.Fprintf(w, "%s%s Alias: %s\n", prefix, aliasPrefix, builder.StringsInterner.MustLookup(importItem.ModuleAlias))
+			}
+
+			if hasOne {
+				currentField++
+				onePrefix := "├─"
+				if currentField == fieldsCount {
+					onePrefix = "└─"
+				}
+				fmt.Fprintf(w, "%s%s One: %s", prefix, onePrefix, formatImportOne(importItem.One, builder))
+				if importItem.One.Alias != 0 {
+					fmt.Fprintf(w, " as %s", builder.StringsInterner.MustLookup(importItem.One.Alias))
+				}
+				fmt.Fprintf(w, "\n")
+			}
+
+			if hasGroup {
+				fmt.Fprintf(w, "%s└─ Group:\n", prefix)
+				for i, pair := range importItem.Group {
+					isLastInGroup := i == len(importItem.Group)-1
+					groupItemPrefix := "├─"
+					if isLastInGroup {
+						groupItemPrefix = "└─"
+					}
+					fmt.Fprintf(w, "%s   %s [%d] %s", prefix, groupItemPrefix, i, builder.StringsInterner.MustLookup(pair.Name))
+					if pair.Alias != 0 {
+						fmt.Fprintf(w, " as %s", builder.StringsInterner.MustLookup(pair.Alias))
+					}
+					fmt.Fprintf(w, "\n")
+				}
+			}
+		}
+	case ast.ItemLet:
+		if letItem, ok := builder.Items.Let(itemID); ok {
+			fields := []struct {
+				label string
+				value string
+				show  bool
+			}{
+				{"Name", builder.StringsInterner.MustLookup(letItem.Name), true},
+				{"Mutable", fmt.Sprintf("%v", letItem.IsMut), true},
+				{"Type", formatTypeExprInline(builder, letItem.Type), letItem.Type.IsValid()},
+				{"Value", formatExprSummary(builder, letItem.Value), true},
+			}
+
+			visible := 0
+			for _, f := range fields {
+				if f.show {
+					visible++
+				}
+			}
+
+			current := 0
+			for _, f := range fields {
+				if !f.show {
+					continue
+				}
+				current++
+				fieldPrefix := "├─"
+				if current == visible {
+					fieldPrefix = "└─"
+				}
+				fmt.Fprintf(w, "%s%s %s: %s\n", prefix, fieldPrefix, f.label, f.value)
+			}
+		}
+	}
+
+	return nil
+}
+
+func formatItemJSON(builder *ast.Builder, itemID ast.ItemID) (ASTNodeOutput, error) {
+	item := builder.Items.Get(itemID)
+	if item == nil {
+		return ASTNodeOutput{}, fmt.Errorf("item not found")
+	}
+
+	output := ASTNodeOutput{
+		Type: "Item",
+		Kind: formatItemKind(item.Kind),
+		Span: item.Span,
+	}
+
+	// Handle special items with payload
+	switch item.Kind {
+	case ast.ItemImport:
+		if importItem, ok := builder.Items.Import(itemID); ok {
+			fields := make(map[string]any)
+
+			var moduleStrs []string
+			for _, stringID := range importItem.Module {
+				moduleStrs = append(moduleStrs, builder.StringsInterner.MustLookup(stringID))
+			}
+			fields["module"] = moduleStrs
+
+			if importItem.ModuleAlias != 0 {
+				fields["moduleAlias"] = builder.StringsInterner.MustLookup(importItem.ModuleAlias)
+			}
+
+			if importItem.HasOne {
+				oneMap := map[string]any{
+					"name": formatImportOne(importItem.One, builder),
+				}
+				if importItem.One.Alias != 0 {
+					oneMap["alias"] = builder.StringsInterner.MustLookup(importItem.One.Alias)
+				}
+				fields["one"] = oneMap
+			}
+
+			if len(importItem.Group) > 0 {
+				var groupItems []map[string]any
+				for _, pair := range importItem.Group {
+					pairMap := map[string]any{
+						"name": builder.StringsInterner.MustLookup(pair.Name),
+					}
+					if pair.Alias != 0 {
+						pairMap["alias"] = builder.StringsInterner.MustLookup(pair.Alias)
+					}
+					groupItems = append(groupItems, pairMap)
+				}
+				fields["group"] = groupItems
+			}
+
+			output.Fields = fields
+		}
+	case ast.ItemLet:
+		if letItem, ok := builder.Items.Let(itemID); ok {
+			fields := map[string]any{
+				"name":     builder.StringsInterner.MustLookup(letItem.Name),
+				"isMut":    letItem.IsMut,
+				"value":    formatExprInline(builder, letItem.Value),
+				"valueSet": letItem.Value.IsValid(),
+				"type":     formatTypeExprInline(builder, letItem.Type),
+				"typeSet":  letItem.Type.IsValid(),
+			}
+			if letItem.Value.IsValid() {
+				fields["valueExprID"] = uint32(letItem.Value)
+			}
+			output.Fields = fields
+		}
+	}
+
+	return output, nil
+}
+
+func formatItemKind(kind ast.ItemKind) string {
+	switch kind {
+	case ast.ItemFn:
+		return "Fn"
+	case ast.ItemLet:
+		return "Let"
+	case ast.ItemType:
+		return "Type"
+	case ast.ItemNewtype:
+		return "Newtype"
+	case ast.ItemAlias:
+		return "Alias"
+	case ast.ItemLiteral:
+		return "Literal"
+	case ast.ItemTag:
+		return "Tag"
+	case ast.ItemExtern:
+		return "Extern"
+	case ast.ItemPragma:
+		return "Pragma"
+	case ast.ItemImport:
+		return "Import"
+	case ast.ItemMacro:
+		return "Macro"
+	default:
+		return fmt.Sprintf("Unknown(%d)", kind)
+	}
+}
+
+func formatImportOne(one ast.ImportOne, builder *ast.Builder) string {
+	if one.Name == 0 {
+		return "*"
+	}
+	return builder.StringsInterner.MustLookup(one.Name)
+}
