+++ internal/diagfmt/ast_tree.go
@@ -1,0 +1,234 @@
+package diagfmt
+
+import (
+	"fmt"
+	"strings"
+	"surge/internal/ast"
+	"surge/internal/source"
+)
+
+type treeNode struct {
+	label    string
+	children []*treeNode
+}
+
+type treeBlock struct {
+	lines []string
+	width int
+	root  int
+}
+
+func buildFileTreeNode(builder *ast.Builder, fileID ast.FileID, fs *source.FileSet) *treeNode {
+	file := builder.Files.Get(fileID)
+	header := "File"
+	if fs != nil {
+		srcFile := fs.Get(file.Span.File)
+		header = srcFile.FormatPath("auto", fs.BaseDir())
+	}
+	root := &treeNode{
+		label: fmt.Sprintf("%s (span: %s)", header, formatSpan(file.Span, fs)),
+	}
+
+	for idx, itemID := range file.Items {
+		root.children = append(root.children, buildItemTreeNode(builder, itemID, fs, idx))
+	}
+
+	return root
+}
+
+func buildItemTreeNode(builder *ast.Builder, itemID ast.ItemID, fs *source.FileSet, idx int) *treeNode {
+	item := builder.Items.Get(itemID)
+	if item == nil {
+		return &treeNode{label: fmt.Sprintf("Item[%d]: <nil>", idx)}
+	}
+
+	node := &treeNode{
+		label: fmt.Sprintf("Item[%d]: %s (span: %s)", idx, formatItemKind(item.Kind), formatSpan(item.Span, fs)),
+	}
+
+	switch item.Kind {
+	case ast.ItemImport:
+		if importItem, ok := builder.Items.Import(itemID); ok {
+			moduleNode := &treeNode{label: "Module"}
+			for _, stringID := range importItem.Module {
+				segment := builder.StringsInterner.MustLookup(stringID)
+				moduleNode.children = append(moduleNode.children, &treeNode{label: segment})
+			}
+			node.children = append(node.children, moduleNode)
+
+			if importItem.ModuleAlias != 0 {
+				alias := builder.StringsInterner.MustLookup(importItem.ModuleAlias)
+				node.children = append(node.children, &treeNode{label: fmt.Sprintf("Alias: %s", alias)})
+			}
+
+			if importItem.HasOne {
+				label := fmt.Sprintf("One: %s", formatImportOne(importItem.One, builder))
+				if importItem.One.Alias != 0 {
+					alias := builder.StringsInterner.MustLookup(importItem.One.Alias)
+					label = fmt.Sprintf("%s as %s", label, alias)
+				}
+				node.children = append(node.children, &treeNode{label: label})
+			}
+
+			if len(importItem.Group) > 0 {
+				groupNode := &treeNode{label: "Group"}
+				for i, pair := range importItem.Group {
+					name := builder.StringsInterner.MustLookup(pair.Name)
+					entry := fmt.Sprintf("[%d] %s", i, name)
+					if pair.Alias != 0 {
+						entry = fmt.Sprintf("%s as %s", entry, builder.StringsInterner.MustLookup(pair.Alias))
+					}
+					groupNode.children = append(groupNode.children, &treeNode{label: entry})
+				}
+				node.children = append(node.children, groupNode)
+			}
+		}
+	case ast.ItemLet:
+		if letItem, ok := builder.Items.Let(itemID); ok {
+			nameNode := &treeNode{label: fmt.Sprintf("Name: %s", builder.StringsInterner.MustLookup(letItem.Name))}
+			mutNode := &treeNode{label: fmt.Sprintf("Mutable: %v", letItem.IsMut)}
+			node.children = append(node.children, nameNode, mutNode)
+
+			if letItem.Type.IsValid() {
+				typeNode := &treeNode{label: fmt.Sprintf("Type: %s", formatTypeExprInline(builder, letItem.Type))}
+				node.children = append(node.children, typeNode)
+			}
+
+			valueLabel := fmt.Sprintf("Value: %s", formatExprSummary(builder, letItem.Value))
+			node.children = append(node.children, &treeNode{label: valueLabel})
+		}
+	}
+
+	return node
+}
+
+func renderTree(node *treeNode) treeBlock {
+	label := node.label
+	labelWidth := len(label)
+
+	if len(node.children) == 0 {
+		return treeBlock{
+			lines: []string{label},
+			width: labelWidth,
+			root:  labelWidth / 2,
+		}
+	}
+
+	childBlocks := make([]treeBlock, len(node.children))
+	maxChildHeight := 0
+	for i, child := range node.children {
+		childBlocks[i] = renderTree(child)
+		if len(childBlocks[i].lines) > maxChildHeight {
+			maxChildHeight = len(childBlocks[i].lines)
+		}
+	}
+
+	const spacing = 3
+
+	positions := make([]int, len(childBlocks))
+	totalWidth := 0
+	for i, block := range childBlocks {
+		positions[i] = totalWidth + block.root
+		totalWidth += block.width
+		if i != len(childBlocks)-1 {
+			totalWidth += spacing
+		}
+	}
+
+	childrenCenter := (positions[0] + positions[len(positions)-1]) / 2
+	rootPos := labelWidth / 2
+	shift := childrenCenter - rootPos
+
+	childPrefix := 0
+	if shift < 0 {
+		childPrefix = -shift
+		for i := range positions {
+			positions[i] += childPrefix
+		}
+		totalWidth += childPrefix
+		shift = 0
+		rootPos = labelWidth / 2
+	} else {
+		rootPos += shift
+	}
+
+	width := totalWidth
+	rootLine := label
+	if shift > 0 {
+		rootLine = strings.Repeat(" ", shift) + label
+	}
+	if len(rootLine) < width {
+		rootLine += strings.Repeat(" ", width-len(rootLine))
+	} else if len(rootLine) > width {
+		width = len(rootLine)
+		for i := range positions {
+			if positions[i] >= width {
+				width = positions[i] + 1
+			}
+		}
+		if len(rootLine) < width {
+			rootLine += strings.Repeat(" ", width-len(rootLine))
+		}
+	}
+
+	connector := make([]byte, width)
+	for i := range connector {
+		connector[i] = ' '
+	}
+	if rootPos >= width {
+		needed := rootPos - width + 1
+		rootLine += strings.Repeat(" ", needed)
+		connector = append(connector, make([]byte, needed)...)
+		for i := width; i < len(connector); i++ {
+			connector[i] = ' '
+		}
+		width = len(connector)
+	}
+	connector[rootPos] = '|'
+	for _, pos := range positions {
+		if pos < rootPos {
+			connector[pos] = '/'
+		} else if pos > rootPos {
+			connector[pos] = '\\'
+		} else {
+			connector[pos] = '|'
+		}
+	}
+	connectorLine := string(connector)
+
+	childLines := make([]string, maxChildHeight)
+	for row := 0; row < maxChildHeight; row++ {
+		var sb strings.Builder
+		if childPrefix > 0 {
+			sb.WriteString(strings.Repeat(" ", childPrefix))
+		}
+		for i, block := range childBlocks {
+			line := ""
+			if row < len(block.lines) {
+				line = block.lines[row]
+			}
+			if len(line) < block.width {
+				line += strings.Repeat(" ", block.width-len(line))
+			}
+			sb.WriteString(line)
+			if i != len(childBlocks)-1 {
+				sb.WriteString(strings.Repeat(" ", spacing))
+			}
+		}
+		rowStr := sb.String()
+		if len(rowStr) < width {
+			rowStr += strings.Repeat(" ", width-len(rowStr))
+		}
+		childLines[row] = rowStr
+	}
+
+	lines := make([]string, 0, 2+len(childLines))
+	lines = append(lines, rootLine, connectorLine)
+	lines = append(lines, childLines...)
+
+	return treeBlock{
+		lines: lines,
+		width: width,
+		root:  rootPos,
+	}
+}
